Image importers - implemented requested changes
+ renamed Jp2/jpeg importers to be more related to their implementation
+ added diamond operators where available
+ reorganized imports in new importers
+ consumerRegistery in Import Process now returned as unmodiffiableList
